+++ misc/dbt-materialize/dbt/adapters/materialize/relation.py
@@ -17,9 +17,10 @@
 from dataclasses import dataclass
 from typing import Optional, Type
 
-from dbt.adapters.postgres import PostgresRelation
-from dbt.dataclass_schema import StrEnum
-from dbt.utils import classproperty
+from dbt_common.dataclass_schema import StrEnum
+
+from dbt.adapters.postgres.relation import PostgresRelation
+from dbt.adapters.utils import classproperty
 
 
 # types in ./misc/dbt-materialize need to import generic types from typing