+++ misc/dbt-materialize/tests/adapter/test_empty.py
@@ -1,4 +1,3 @@
-# Copyright 2020 Josh Wills. All rights reserved.
 # Copyright Materialize, Inc. and contributors. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,5 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# If you bump this version, bump it in setup.py too.
-version = "1.8.0"
+from dbt.tests.adapter.empty.test_empty import BaseTestEmpty
+
+
+class TestMaterializeEmpty(BaseTestEmpty):
+    pass